refactor: custom hook을 적용하여 state와 로직 분리
